requirements.txt
tensorflow==2.8.0
keras
numpy
Pillow==8.4.0
google-auth
google-auth-httplib2
google-api-python-client
torch
flask
torch-vision 